feat: installed bottom tabs
