+++ App.js
@@ -1,13 +1,32 @@
+import { Text } from "react-native";
 import { StatusBar as ExpoStatusBar } from "expo-status-bar";
 import { RestaurantScreen } from "./src/features/restaurants/screens/restaurants.screen";
 import { ThemeProvider } from "styled-components/native";
-import { theme } from "./src/infrastructure/theme";
+import { NavigationContainer } from "@react-navigation/native";
+import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
 
 import {
   useFonts as useOswald,
   Oswald_400Regular,
 } from "@expo-google-fonts/oswald";
 import { useFonts as useLato, Lato_400Regular } from "@expo-google-fonts/lato";
+import { SafeArea } from "./src/components/utility/safe-area.component";
+import { theme } from "./src/infrastructure/theme";
+
+const SettingsScreen = () => {
+  return (
+    <SafeArea>
+      <Text>Settings Screen!</Text>
+    </SafeArea>
+  );
+};
+const MapScreen = () => {
+  return (
+    <SafeArea>
+      <Text>Map Screen!</Text>
+    </SafeArea>
+  );
+};
 
 export default function App() {
   const [oswaldLoaded] = useOswald({
@@ -21,10 +40,18 @@ export default function App() {
     return null;
   }
 
+  const Tab = createBottomTabNavigator();
+
   return (
     <>
       <ThemeProvider theme={theme}>
-        <RestaurantScreen />
+        <NavigationContainer>
+          <Tab.Navigator>
+            <Tab.Screen name="Home" component={RestaurantScreen} />
+            <Tab.Screen name="Map" component={MapScreen} />
+            <Tab.Screen name="Settings" component={SettingsScreen} />
+          </Tab.Navigator>
+        </NavigationContainer>
         <ExpoStatusBar style="auto" />
       </ThemeProvider>
     </>
