+++ src/components/utility/safe-area.component.js
@@ -1,0 +1,8 @@
+import { SafeAreaView, StatusBar } from "react-native";
+
+import styled from "styled-components/native";
+
+export const SafeArea = styled(SafeAreaView)`
+  flex: 1;
+  margin-top: ${StatusBar.currentHeight}px;
+`;
